+++ src/main/java/com/anesuandtatenda/studentlecturerplatform/service/TimeSlotLecturerService.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public interface TimeSlotLecturerService extends BaseService<TimeSlotLecturer, TimeSlotLecturerRequest, TimeSlotLecturerRequest> {
 
-    List<TimeSlotLecturer> findByDayOfWeek(int dayOfWeek);
+    List<TimeSlotLecturer> findByDayOfWeek(long dayOfWeek);
 
     List<TimeSlotLecturer> findByAccountId(long id);
 
-    List<TimeSlotLecturer> findByDayOfWeekAndAccountId(int dayOfWeek, long id);
+    List<TimeSlotLecturer> findByDayOfWeekAndAccountId(long dayOfWeek, long id);
 
-   // TimeSlotLecturer findByDayOfWeekAndAccountIdAAndTimeSlotsId(int dayOfWeek, long id, long timeSlotId);
+   // TimeSlotLecturer findByDayOfWeekAndAccountIdAAndTimeSlotsId(long dayOfWeek, long id, long timeSlotId);
 
 }
